relocated and removed legacy kpi definitions
